requirements.txt
Flask
pyjwt
python-dotenv
requests
